requirements.txt
attrs==21.2.0
Automat==20.2.0
Brotli==1.0.9
cffi==1.15.0
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.0
cssselect==1.1.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.2
Flask-Compress==1.10.1
greenlet==1.1.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
lxml==4.6.4
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
numpy==1.21.5
pandas==1.3.5
parsel==1.6.0
plotly==5.5.0
priority==1.3.0
Protego==0.1.16
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pymongo==4.0.1
PyMySQL==1.0.2
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
queuelib==1.6.2
Scrapy==2.5.1
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.29
tenacity==8.0.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing-extensions==4.0.1
w3lib==1.22.0
Werkzeug==2.0.2
zope.interface==5.4.0
